api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile api/pyproject.toml -o api/requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==5.0.0
    # via passlib
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==46.0.3
    # via python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.121.3
    # via built-frontend (api/pyproject.toml)
fastapi-cli==0.0.16
    # via fastapi
fastapi-cloud-cli==0.5.1
    # via fastapi-cli
fastar==0.6.0
    # via fastapi-cloud-cli
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.11.4
    # via fastapi
passlib==1.7.4
    # via built-frontend (api/pyproject.toml)
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.10.6
    # via fastapi
pydantic-settings==2.12.0
    # via fastapi
pygments==2.19.2
    # via rich
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via built-frontend (api/pyproject.toml)
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
rsa==4.9.1
    # via python-jose
sentry-sdk==2.45.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via sqlmodel
sqlmodel==0.0.27
    # via built-frontend (api/pyproject.toml)
starlette==0.50.0
    # via fastapi
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing==3.10.0.0
    # via built-frontend (api/pyproject.toml)
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via sentry-sdk
uuid==1.30
    # via built-frontend (api/pyproject.toml)
uvicorn==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
